Adding the ability to see updates Asynchronously
- [x] Adding a contrived example where we allow the update to the db record occur from a Celery task.
