+++ dna/api/external_api.py
@@ -3,16 +3,14 @@ import os
 from supabase import create_client, Client
 
 from fastapi import FastAPI, HTTPException
-from fastapi.params import Depends
-from sqlalchemy.orm import Session
-from datetime import datetime
+from fastapi.encoders import jsonable_encoder
 
-from dna.common.model import Profile
-from dna.common.model.session_management import managed_session
+
+from dna.common.data_access.profile_operations import save_or_update_profile
 from dna.common.schemas import ProfileUpdate
 
 
-from dna.tasks.external_tasks import show_message
+from dna.tasks.external_tasks import show_message, update_profile_task
 
 ####
 # TODO: Refactor this such that it uses our configuration
@@ -59,7 +57,7 @@ async def create_profile(profile: dict):
 
 
 @app.put("/profiles/{profile_id}")
-def update_profile(profile_id: str, profile: ProfileUpdate):
+async def update_profile(profile_id: str, profile: ProfileUpdate):
     # Get the profile with the specified ID from the database
     db_profile = save_or_update_profile(profile, profile_id)
 
@@ -67,24 +65,9 @@ def update_profile(profile_id: str, profile: ProfileUpdate):
     return db_profile
 
 
-@managed_session
-def save_or_update_profile(profile, profile_id, session=None):
-    db_profile = session.query(Profile).filter(Profile.id == profile_id).first()
-    # If the profile doesn't exist, return a 404 response
-    if not db_profile:
-        raise HTTPException(status_code=404, detail="Profile not found")
-
-    # Update the profile with the new values
-    if profile.username is not None:
-        db_profile.username = profile.username
-    if profile.full_name is not None:
-        db_profile.full_name = profile.full_name
-    if profile.avatar_url is not None:
-        db_profile.avatar_url = profile.avatar_url
-    if profile.website is not None:
-        db_profile.website = profile.website
-    db_profile.update_at = datetime.now()
-
-    # Commit the changes to the database
-    session.commit()
-    return db_profile
+@app.put("/profiles/async/{profile_id}")
+async def update_profile_async(profile_id: str, profile: ProfileUpdate):
+    # Get the profile with the specified ID from the database
+    update_profile_task.delay(profile_id, jsonable_encoder(profile))
+    # Return the updated profile
+    return {"message": "Profile is being updated asynchronously"}
